Move some ui functions to ui-utils
