+++ packages/rocketchat-ui-utils/client/lib/RoomManager.js
@@ -5,6 +5,7 @@ import { Blaze } from 'meteor/blaze';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { Template } from 'meteor/templating';
 import { roomTypes as _roomTypes } from 'meteor/rocketchat:utils';
+import { fireGlobalEvent } from 'meteor/rocketchat:ui-utils';
 import { promises } from 'meteor/rocketchat:promises';
 import { callbacks } from 'meteor/rocketchat:callbacks';
 import { Notifications } from 'meteor/rocketchat:notifications';
@@ -77,7 +78,7 @@ export const RoomManager = new function() {
 
 										callbacks.run('streamMessage', msg);
 
-										return window.fireGlobalEvent('new-message', msg);
+										return fireGlobalEvent('new-message', msg);
 									}
 								})
 							);
