+++ packages/rocketchat-ui-utils/client/lib/openedRoom.js
@@ -1,0 +1,1 @@
+export let openedRoom; //eslint-disable-line
