+++ packages/rocketchat-ui-utils/package.js
@@ -20,6 +20,8 @@ Package.onUse(function(api) {
 		'rocketchat:streamer',
 		'rocketchat:models',
 		'rocketchat:lazy-load',
+		'rocketchat:emoji',
 	]);
 	api.mainModule('client/index.js', 'client');
+	api.mainModule('server/index.js', 'server');
 });
