+++ packages/rocketchat-ui-utils/server/index.js
@@ -1,0 +1,2 @@
+export { messageProperties } from '../lib/MessageProperties';
+export { MessageTypes } from '../lib/MessageTypes';
